@@ -18,7 +18,7 @@ description: 香月手机APP
 version: 1.1.2+2
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -27,7 +27,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-
+  auto_orientation_v2: ^2.3.6
+  device_info_plus: ^12.1.0
+  flutter_blue_classic: ^0.0.6
+  permission_handler: ^12.0.1
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -36,27 +39,27 @@ dev_dependencies:
     sdk: flutter
 
   # 用dio进行网络接口调用：https://pub.flutter-io.cn/packages/dio
-  dio: ^3.0.10
-  dio_retry: ^0.1.9-beta
+  dio: ^5.4.0
+  dio_smart_retry: ^5.0.0
   # 本地存储：https://pub.flutter-io.cn/packages/shared_preferences
   shared_preferences: ^2.0.3
   # Toast https://pub.flutter-io.cn/packages/fluttertoast    https://github.com/PonnamKarthik/FlutterToast
   fluttertoast: ^8.0.8
   # device_info: 获取设备信息：https://pub.flutter-io.cn/packages/device_info#-readme-tab-
-  device_info: ^2.0.3
-  auto_orientation: ^2.0.2 
+#  device_info: ^2.0.3
+
   # just_audio: ^0.6.14+1
-  flutter_rating_bar: ^3.2.0+1
+  flutter_rating_bar: ^4.0.1
 #  flutter_swiper: ^1.1.6
   flustars: ^2.0.0
-  cached_network_image: ^2.2.0+1
-  event_bus: ^1.1.1
+  cached_network_image: ^3.4.1
+  event_bus: ^2.0.0
   simple_animations: ^3.1.1
-  provider: ^4.3.0
+  provider: ^6.1.2
   # permission_handler: ^2.1.2  权限检查、请求   https://pub.flutter-io.cn/packages/permission_handler#-installing-tab-
-  permission_handler: ^8.2.4 #4.4.0+hotfix.4 # 升级到5.0.0+hotfix.5报错
+#  permission_handler: ^8.2.4 #4.4.0+hotfix.4 # 升级到5.0.0+hotfix.5报错
   flutter_xlider: ^3.1.1 #定制scrollbar
-  app_settings: ^4.1.1
+  app_settings: ^6.1.1
   flutter_bluetooth_serial:
    path: ./flutter_bluetooth_serial-0.4.0/
   
