@@ -1,3 +1,3 @@
 import 'package:integration_test/integration_test_driver.dart';
 
-Future<void> main() => integrationDriver();
+// Future<void> main() => integrationDriver();
